crypto-tracker/backend/requirements.txt
fastapi>=0.68.0
uvicorn>=0.15.0
python-binance>=1.0.19
requests>=2.26.0
transformers>=4.12.0
python-dotenv>=0.19.0
aiohttp>=3.7.4
